Params handling, scaffold for call to the RunStrategyHandler and onSuccess event

+++ src/main/java/com/mns/mojoinvest/server/engine/portfolio/SimplePortfolio.java
@@ -52,8 +52,8 @@ public class SimplePortfolio implements Portfolio {
     public SimplePortfolio(QuoteDao quoteDao, @Assisted PortfolioParams params) {
         this.quoteDao = quoteDao;
         this.positions = new HashMap<Fund, Position>();
-        this.cash = params.getInitialInvestment();
-        this.transactionCost = params.getTransactionCost();
+        this.cash = BigDecimal.valueOf(params.getInitialInvestment());
+        this.transactionCost = BigDecimal.valueOf(params.getTransactionCost());
     }
 
     @Override
